fix test against max send tokens
This commit was SVN r6531.

+++ ompi/mca/btl/gm/btl_gm_frag.h
@@ -109,7 +109,7 @@ OBJ_CLASS_DECLARATION(mca_btl_gm_frag_user_t);
                                                                                                        
 #define MCA_BTL_GM_FRAG_POST(btl,frag) \
 do { \
-    if(opal_list_get_size(&btl->gm_repost) < btl->gm_num_repost) { \
+    if(opal_list_get_size(&btl->gm_repost) < (size_t)btl->gm_num_repost) { \
         OPAL_THREAD_LOCK(&btl->gm_lock);  \
         opal_list_append(&btl->gm_repost, (opal_list_item_t*)frag); \
         OPAL_THREAD_UNLOCK(&btl->gm_lock);  \
